+++ DesktopApplicationTemplate.UI/Views/Shared/ServiceEditorView.xaml.cs
@@ -1,3 +1,4 @@
+using System.Windows;
 using System.Windows.Controls;
 
 namespace DesktopApplicationTemplate.UI.Views;
@@ -8,4 +9,31 @@ public partial class ServiceEditorView : Page
     {
         InitializeComponent();
     }
+
+    public string AdvancedAutomationName
+    {
+        get => (string)GetValue(AdvancedAutomationNameProperty);
+        set => SetValue(AdvancedAutomationNameProperty, value);
+    }
+
+    public static readonly DependencyProperty AdvancedAutomationNameProperty =
+        DependencyProperty.Register(nameof(AdvancedAutomationName), typeof(string), typeof(ServiceEditorView), new PropertyMetadata("Open Advanced Configuration"));
+
+    public string SaveAutomationName
+    {
+        get => (string)GetValue(SaveAutomationNameProperty);
+        set => SetValue(SaveAutomationNameProperty, value);
+    }
+
+    public static readonly DependencyProperty SaveAutomationNameProperty =
+        DependencyProperty.Register(nameof(SaveAutomationName), typeof(string), typeof(ServiceEditorView), new PropertyMetadata("Save Service"));
+
+    public string CancelAutomationName
+    {
+        get => (string)GetValue(CancelAutomationNameProperty);
+        set => SetValue(CancelAutomationNameProperty, value);
+    }
+
+    public static readonly DependencyProperty CancelAutomationNameProperty =
+        DependencyProperty.Register(nameof(CancelAutomationName), typeof(string), typeof(ServiceEditorView), new PropertyMetadata("Cancel"));
 }